systemd: Move functions that only affect metadata to parse time.
The functions that don't have dependencies on the data in ${D} should
run at parse time. This way overrides behave correctly and the
python functions can actually be replaced/redefined in individual
recipes.

i.e:
systemd_postinst_${PN} () {
}

This is useful for times when you don't want to enable or disable a
service automatically on install, but still want the rest of the
functionality of the systemd.bbclass.

Signed-off-by: Randy Witt <rewitt@declaratino.com>